Handle blank lines in config
If the final line of the config had a newline (which is normal on Mac and
Linux), VST treated it as an I/O error.  Fixed.

If any line consisted only of `\r`, which is what a blank line in a
CRLF-delimited file looks like on Mac or Linux, VST treated it as a syntax
error.  More generally, every line read had a `\r` on the end that had to
be stripped to parse out the correct RHS.  Fixed.